Log to stdout by default.
`mysqld_safe` already logs to stdout, but MySQL error logs do not.
They default to `/var/lib/mysql/mysql-0.err`, where `mysql-0` is the hostname.

This change makes it default to stdout and is configurable.

I haven't tested this; it's not immediately clear how (clone code, push to a private GitLab fork, turn on container registry, build, push to container registry, tell Flight Deck Cluster to pull that?). It's kind of...not worth it for a one-line change—I can always `exec` in and check logs that way—but figured I'd PR this and see where it went or if you randomly had time.